+++ app/controllers/anomonitor/application_controller.rb
@@ -6,5 +6,16 @@ module Anomonitor
     layout "anomonitor/application"
 
     helper Anomonitor::Engine.helpers
+
+    before_action :authenticate_anomonitor!
+
+    private
+
+    def authenticate_anomonitor!
+      auth = Anomonitor.config.authenticate
+      return if auth.nil?
+
+      instance_exec(&auth)
+    end
   end
 end